feat(elements): move table to separate file
